Add task class
