--- src/main/java/duke/FileReading.java
@@ -1,23 +1,0 @@
-package duke;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-
-public class FileReading {
-    private static void writeToFile(String filePath, String textToAdd) throws IOException {
-        FileWriter fw = new FileWriter(filePath);
-        fw.write(textToAdd);
-        fw.close();
-    }
-
-    public static void writeToPath(String filePath, String textToAdd) {
-        String file = filePath;
-        try {
-            writeToFile(file, textToAdd);
-        } catch (IOException e) {
-            System.out.println("Something went wrong: " + e.getMessage());
-        }
-    }
-}
